Fixed generation for MSSQL.

+++ src/main/resources/ddl/results/pathway_analysis_events.sql
@@ -2,8 +2,8 @@ IF OBJECT_ID('@results_schema.pathway_analysis_events', 'U') IS NULL
 CREATE TABLE @results_schema.pathway_analysis_events
 (
 	pathway_analysis_generation_id BIGINT NOT NULL,
-	combo_id INTEGER NOT NULL,
-	subject_id INTEGER NOT NULL,
+	combo_id BIGINT NOT NULL,
+	subject_id BIGINT NOT NULL,
 	cohort_start_date DATETIME NOT NULL,
 	cohort_end_date DATETIME NOT NULL
 );
